Add files via upload
Added description for the remaining arguments. Removed noiseInfer as an argument (as we have decided to only allow fixed sd of the observational noise for the time being). Added explicitNoiseInfer as an argument (to allow the sd of the observational noise to be inferred ONLY when the explicit solution is invoked).