Filled in the missing parts of vignette......
Filled in the missing parts of vignette, changed some wording in some of the sentences, changed the species labels to be x_1 and x_2 etc instead of x and y (this was to avoid confusion with previous notation of y for data. It also helps identify more readily where each species appears in the code).

If there is anything you want changed, please let me know. Sorry about the delay, just got my replacement computer today.