Changed dataConstants to work with observedVariables
Updated to allow the code to run with missing species. The previous calculation of dataConstants caused the predictions from the ODEs to be NA when dealing with missing species. The calculation of the likelihood could then not be implemented and the program crashed. Now dataConstants is calculated for the observed species.